% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warmup.R
\name{warmup}
\alias{warmup}
\title{Perform warmup for a Stan model}
\usage{
warmup(
  file,
  stan_fit,
  window_size = 100,
  max_num_windows = 10,
  chains = 4,
  target_rhat = 1.05,
  target_ess = 50,
  print_stdout = FALSE,
  ...
)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}

\item{stan_fit}{PARAM_DESCRIPTION}

\item{window_size}{PARAM_DESCRIPTION, Default: 100}

\item{max_num_windows}{PARAM_DESCRIPTION, Default: 10}

\item{chains}{PARAM_DESCRIPTION, Default: 4}

\item{target_rhat}{PARAM_DESCRIPTION, Default: 1.05}

\item{target_ess}{PARAM_DESCRIPTION, Default: 50}

\item{print_stdout}{PARAM_DESCRIPTION, Default: FALSE}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
